+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.4.0...@synapsecns/synapse-interface@0.5.0) (2024-03-08)
+
+
+### Features
+
+* **synapse-interface:** update button states ([#2214](https://github.com/synapsecns/sanguine/issues/2214)) ([90df2bf](https://github.com/synapsecns/sanguine/commit/90df2bf462e74f4c3e492791e26559b9ee5bc72f))
+
+
+
+
+
+# [0.4.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.3.0...@synapsecns/synapse-interface@0.4.0) (2024-03-08)
+
+
+### Features
+
+* **synapse-interface:** post transaction balances ([#2079](https://github.com/synapsecns/sanguine/issues/2079)) ([2a95040](https://github.com/synapsecns/sanguine/commit/2a95040c7a6535d21e004b4bf4bcb7d022a9cda7))
+
+
+
+
+
+# [0.3.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.2.2...@synapsecns/synapse-interface@0.3.0) (2024-03-07)
+
+
+### Features
+
+* **synapse-interface:** transaction status ([#2052](https://github.com/synapsecns/sanguine/issues/2052)) ([e5ae5a2](https://github.com/synapsecns/sanguine/commit/e5ae5a299acb8d3e2ed12d960aae684b473aba6d))
+
+
+
+
+
 ## [0.2.2](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.2.1...@synapsecns/synapse-interface@0.2.2) (2024-03-07)
 
 **Note:** Version bump only for package @synapsecns/synapse-interface